refactor: remove `/rmrk` with Pagination
