+++ components/rmrk/utils.ts
@@ -41,9 +41,3 @@ export const resolveMedia = (mimeType?: string): MediaType => {
 
   return result
 }
-
-export const getRandomIntInRange = (min: number, max: number): number => {
-  min = Math.ceil(min)
-  max = Math.floor(max)
-  return Math.floor(Math.random() * (max - min + 1)) + min
-}
